refactoring

+++ services/sangraha-service/src/main/java/sm/selflearn/samskrtam/sangraha/service/LlmProperties.java
@@ -20,5 +20,5 @@ public class LlmProperties {
      * Если null — используется дефолт провайдера.
      * Рекомендуется 8192 для длинных стихов с two-pass.
      */
-    private Long maxCompletionTokens;
+    private Integer maxCompletionTokens;
 }
